refactor: whatsapp related methods
