+++ src/manager.ts
@@ -20,7 +20,6 @@ import {
 } from './utils.js'
 import { RequestManager } from './request/requestManager.js'
 import { CacheManager } from './data-manager/cache-manager.js'
-import { getWhatsApp } from './whatsapp.js'
 
 import type { PuppetWhatsAppOptions } from './puppet-whatsapp.js'
 import type {
@@ -33,24 +32,24 @@ import type {
 } from './schema/whatsapp-type.js'
 import ScheduleManager from './schedule/schedule-manager.js'
 import type { ManagerEvents } from './manager-event.js'
-import WhatsAppEvent from './whatsapp/whatsapp-events.js'
+import WhatsAppManager from './whatsapp/whatsapp-manager.js'
 
 const logger = withPrefix(`${PRE} Manager`)
 export class Manager extends EE<ManagerEvents> {
 
   whatsAppClient?: WhatsAppClientType
-  whatsAppEvent: WhatsAppEvent
-  private _requestManager?: RequestManager
+  whatsAppManager: WhatsAppManager
+  _requestManager?: RequestManager
   cacheManager?: CacheManager
   scheduleManager: ScheduleManager
-  botId?: string
-  fetchingMessages: boolean = false
-  loadingData: boolean = false
+
+  private fetchingMessages: boolean = false
+  private heartbeatTimer?: NodeJS.Timer
 
   constructor (private options: PuppetWhatsAppOptions) {
     super()
     this.scheduleManager = new ScheduleManager(this)
-    this.whatsAppEvent = new WhatsAppEvent(this)
+    this.whatsAppManager = new WhatsAppManager(this)
   }
 
   public getOptions () {
@@ -63,7 +62,7 @@ export class Manager extends EE<ManagerEvents> {
 
   public async start (session?: ClientSession) {
     logger.info('start()')
-    this.whatsAppClient = await getWhatsApp(this.options['puppeteerOptions'], session)
+    this.whatsAppClient = await this.whatsAppManager.initWhatsAppClient(this.options['puppeteerOptions'], session)
     this.whatsAppClient
       .initialize()
       .then(() => logger.verbose('start() whatsapp.initialize() done.'))
@@ -75,8 +74,8 @@ export class Manager extends EE<ManagerEvents> {
         }
       })
 
-    await this.whatsAppEvent.initWhatsAppEvents(this.whatsAppClient)
     this._requestManager = new RequestManager(this.whatsAppClient)
+    await this.whatsAppManager.initWhatsAppEvents(this.whatsAppClient)
 
     this.startHeartbeat()
     return this.whatsAppClient
@@ -90,13 +89,13 @@ export class Manager extends EE<ManagerEvents> {
     }
     await this.releaseCache()
     this._requestManager = undefined
-    this.resetAllVarInMemory()
+    this.whatsAppManager.clearWhatsAppRelatedData()
 
     this.stopHeartbeat()
   }
 
   public async syncContactOrRoomList () {
-    const whatsapp = this.getWhatsApp()
+    const whatsapp = this.getWhatsAppClient()
     const contactList: WhatsAppContact[] = await whatsapp.getContacts()
     const contactOrRoomList = contactList.filter(c => c.id.server !== 'broadcast' && c.id._serialized !== '0@c.us')
     return contactOrRoomList
@@ -140,13 +139,17 @@ export class Manager extends EE<ManagerEvents> {
       const batchSize = 50
       await batchProcess(batchSize, messageList, async (message: WhatsAppMessage) => {
         if (message.ack === MessageAck.ACK_DEVICE || message.ack === MessageAck.ACK_READ) {
-          await this.whatsAppEvent.onMessage(message)
+          await this.processMessage(message)
         }
       })
     } catch (error) {
       logger.error(`fetchMessages error: ${(error as Error).message}`)
     }
     this.fetchingMessages = false
+  }
+
+  public async processMessage (message: WhatsAppMessage) {
+    await this.whatsAppManager.getMessageEventHandler().onMessage(message)
   }
 
   public async getCacheManager () {
@@ -197,7 +200,7 @@ export class Manager extends EE<ManagerEvents> {
     }
   }
 
-  public getWhatsApp () {
+  public getWhatsAppClient () {
     if (!this.whatsAppClient) {
       throw WAError(WA_ERROR_TYPE.ERR_INIT, 'Not init whatsapp')
     }
@@ -215,14 +218,6 @@ export class Manager extends EE<ManagerEvents> {
     // FIXME: How to deal with pendingParticipants? Maybe we should find which case could has this attribute.
     return roomChat.participants.map(m => m.id._serialized)
   }
-
-  private resetAllVarInMemory () {
-    this.botId = undefined
-    this.loadingData = false
-    this.fetchingMessages = false
-  }
-
-  private heartbeatTimer?: NodeJS.Timer
 
   private startHeartbeat () {
     if (!this.heartbeatTimer) {
@@ -247,7 +242,7 @@ export class Manager extends EE<ManagerEvents> {
      * it will not return true if the Chromium process is terminated with command + q
      */
 
-    const alive = this.getWhatsApp().pupBrowser?.isConnected()
+    const alive = this.getWhatsAppClient().pupBrowser?.isConnected()
     if (alive) {
       this.asystoleCount = 0
       this.emit('heartbeat', 'puppeteer still connected')
